chore: separacao de responsabilidades e ajustes de endpoints
