+++ src/main/java/br/com/cumbuca/controller/AvaliacaoController.java
@@ -2,8 +2,6 @@ package br.com.cumbuca.controller;
 
 import br.com.cumbuca.dto.avaliacao.AvaliacaoRequestDTO;
 import br.com.cumbuca.dto.avaliacao.AvaliacaoResponseDTO;
-import br.com.cumbuca.dto.comentario.ComentarioResponseDTO;
-import br.com.cumbuca.dto.curtida.CurtidaResponseDTO;
 import br.com.cumbuca.service.avaliacao.AvaliacaoService;
 import br.com.cumbuca.service.comentario.ComentarioService;
 import br.com.cumbuca.service.curtida.CurtidaService;
@@ -18,7 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -77,19 +74,6 @@ public class AvaliacaoController {
             @RequestParam(required = false) Long idEstabelecimento) {
         final List<AvaliacaoResponseDTO> avaliacoes = avaliacaoService.listar(idUsuario, idEstabelecimento);
         return ResponseEntity.ok(avaliacoes);
-    }
-
-    @PostMapping("/curtir/{id}")
-    public ResponseEntity<CurtidaResponseDTO> curtir(@PathVariable Long id) {
-        final CurtidaResponseDTO curtida = curtidaService.curtir(id);
-        return ResponseEntity.ok(curtida);
-    }
-
-    @PostMapping("/comentar/{id}")
-    public ResponseEntity<ComentarioResponseDTO> comentar(@PathVariable Long id,
-                                                          @Valid @RequestBody String texto) {
-        final ComentarioResponseDTO comentario = comentarioService.comentar(id, texto);
-        return ResponseEntity.ok(comentario);
     }
 
 }
