+++ src/main/java/br/com/cumbuca/controller/ComentarioController.java
@@ -1,15 +1,17 @@
 package br.com.cumbuca.controller;
 
+import br.com.cumbuca.dto.comentario.ComentarioResponseDTO;
 import br.com.cumbuca.service.comentario.ComentarioService;
+import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 
 @RestController
-@RequestMapping("/comentario")
 public class ComentarioController {
 
     private ComentarioService comentarioService;
@@ -18,7 +20,15 @@ public class ComentarioController {
         this.comentarioService = comentarioService;
     }
 
-    @DeleteMapping("/remover/{id}")
+
+    @PostMapping("avaliacao/comentar/{id}")
+    public ResponseEntity<ComentarioResponseDTO> comentar(@PathVariable Long id,
+                                                          @Valid @RequestBody String texto) {
+        final ComentarioResponseDTO comentario = comentarioService.comentar(id, texto);
+        return ResponseEntity.ok(comentario);
+    }
+
+    @DeleteMapping("comentario/remover/{id}")
     public ResponseEntity<Void> removerComentario(@PathVariable Long id) {
         comentarioService.removerComentario(id);
         return ResponseEntity.noContent().build();
